requirements.txt
h5py
scipy
Pillow
torchvision==0.7.0+cu101
tqdm
torch==1.6.0+cu101
requests
numpy
matplotlib
tensorflow-gpu==1.15
pyyaml
